requirements.txt
streamlit
langchain
langchain-groq
langgraph
langchain-community
pydantic
python-dotenv
duckdb # Using a simple, fast, in-process database